refactor: move worker code to a folder
