+++ packages/yarnpkg-core/sources/ZipWorker/index.d.ts
@@ -1,0 +1,2 @@
+export function getContent(): string;
+export type {ConvertToZipPayload} from './Worker';
